Update type of CoverRate fields

+++ include/xrpl/protocol/Protocol.h
@@ -100,7 +100,7 @@ std::uint16_t constexpr maxFeeRate = 10'000;
     loss capital operations are between 0 and 100000 inclusive. A value of 1 is
     equivalent to 1/10 bps or 0.001%.
 */
-std::uint16_t constexpr maxCoverRate = 100'000;
+std::uint32_t constexpr maxCoverRate = 100'000;
 
 /** The maximum length of a URI inside an NFT */
 std::size_t constexpr maxTokenURILength = 256;
